requirements.txt
streamlit==0.82.0
mols2grid==0.0.6
pyarrow==4.0.1
rdkit-pypi==2021.3.2.3
chembl_webresource_client==0.10.4